updated CLI/Truck files
Started digging into the other files.  I should have read them all first and figured out what went to where before I started on the truck code.  It had a blueprint in the car file and all I would have had to do was expand that idea.

started to work on the CLI file so I can connect them.  updated the file for trucks and motorbikes.